# [Static Runtime] Added NNC implementation for signed log1p kernel.