--- a/benchmarks/static_runtime/deep_wide_pt.h
+++ b/benchmarks/static_runtime/deep_wide_pt.h
@@ -138,3 +138,5 @@ torch::jit::Module getLeakyReLUScriptModel();
 torch::jit::Module getLeakyReLUConstScriptModel();
 
 torch::jit::Module getLongScriptModel();
+
+torch::jit::Module getSignedLog1pModel();